[Build] Scrapping messages
Retrieving full messages for now, since we don't have a user entity so far its still static.

utils:
messageToBackend.ts - added a function to send the messages inputted by our users to the backend.

For the ''user'' location I've implemented it with the browser geolocation API for now, later we can get from the user information